now it works with redux; odd files were removed
