--- task-2/app/app.global.css
@@ -1,14 +1,0 @@
-body {
-	position: relative;
-	color: white;
-	background: #232c39
-		linear-gradient(45deg, rgba(0, 216, 255, 0.5) 10%, rgba(0, 1, 127, 0.7));
-	font-family: Arial, Helvetica, Helvetica Neue, serif;
-	overflow-y: hidden;
-}
-
-#root,
-body {
-	margin: 0;
-	height: 100vh;
-}
