+++ task-2/app/assets/app.global.css
@@ -1,0 +1,5 @@
+#root,
+body {
+	margin: 0;
+	height: 100vh;
+}
