--- task-2/app/components/Counter.js
@@ -1,65 +1,0 @@
-import React, { Component } from 'react';
-import styles from './Counter.css';
-
-type Props = {
-	increment: () => void,
-	incrementIfOdd: () => void,
-	incrementAsync: () => void,
-	decrement: () => void,
-	counter: number
-};
-
-export default class Counter extends Component<Props> {
-	props: Props;
-
-	render() {
-		const {
-			increment,
-			incrementIfOdd,
-			incrementAsync,
-			decrement,
-			counter
-		} = this.props;
-		return (
-			<div>
-				<div className={`counter ${styles.counter}`} data-tid="counter">
-					{counter}
-				</div>
-				<div className={styles.btnGroup}>
-					<button
-						className={styles.btn}
-						onClick={increment}
-						data-tclass="btn"
-						type="button"
-					>
-						+
-					</button>
-					<button
-						className={styles.btn}
-						onClick={decrement}
-						data-tclass="btn"
-						type="button"
-					>
-						-
-					</button>
-					<button
-						className={styles.btn}
-						onClick={incrementIfOdd}
-						data-tclass="btn"
-						type="button"
-					>
-						odd
-					</button>
-					<button
-						className={styles.btn}
-						onClick={() => incrementAsync()}
-						data-tclass="btn"
-						type="button"
-					>
-						async
-					</button>
-				</div>
-			</div>
-		);
-	}
-}
