+++ task-2/app/data/store/configureStore.dev.js
@@ -1,43 +1,30 @@
 import { applyMiddleware, compose, createStore } from 'redux';
-import thunk from 'redux-thunk';
 import { createHashHistory } from 'history';
-import { routerActions, routerMiddleware } from 'connected-react-router';
 import { createLogger } from 'redux-logger';
-import createRootReducer from '../reducers';
-import * as counterActions from '../actions/counter';
-import type { counterStateType } from '../reducers/types';
+import createRootReducer from '../reducer';
+import { addEllipse } from '../modules/canvas';
 
 const history = createHashHistory();
 
 const rootReducer = createRootReducer(history);
 
-const configureStore = (initialState?: counterStateType) => {
+const configureStore = (initialState) => {
 	// Redux Configuration
 	const middleware = [];
 	const enhancers = [];
 
-	// Thunk Middleware
-	middleware.push(thunk);
-
-	// Logging Middleware
 	const logger = createLogger({
 		level: 'info',
 		collapsed: true
 	});
 
-	// Skip redux logs in console during the tests
 	if (process.env.NODE_ENV !== 'test') {
 		middleware.push(logger);
 	}
 
-	// Router Middleware
-	const router = routerMiddleware(history);
-	middleware.push(router);
-
 	// Redux DevTools Configuration
 	const actionCreators = {
-		...counterActions,
-		...routerActions
+		addEllipse,
 	};
 	// If Redux DevTools Extension is installed use it, otherwise use Redux compose
 	/* eslint-disable no-underscore-dangle */
@@ -53,13 +40,12 @@ const configureStore = (initialState?: counterStateType) => {
 	enhancers.push(applyMiddleware(...middleware));
 	const enhancer = composeEnhancers(...enhancers);
 
-	// Create Store
 	const store = createStore(rootReducer, initialState, enhancer);
 
 	if (module.hot) {
 		module.hot.accept(
 			'../reducers',
-			() => store.replaceReducer(require('../reducers')) // eslint-disable-line global-require
+			() => store.replaceReducer(require('../reducer')) // eslint-disable-line global-require
 		);
 	}
