+++ task-2/app/data/store/configureStore.prod.js
@@ -1,17 +1,15 @@
-// @flow
 import { applyMiddleware, createStore } from 'redux';
 import thunk from 'redux-thunk';
 import { createHashHistory } from 'history';
 import { routerMiddleware } from 'connected-react-router';
-import createRootReducer from '../reducers';
-import type { counterStateType } from '../reducers/types';
+import createRootReducer from '../reducer';
 
 const history = createHashHistory();
 const rootReducer = createRootReducer(history);
 const router = routerMiddleware(history);
 const enhancer = applyMiddleware(thunk, router);
 
-function configureStore(initialState?: counterStateType) {
+function configureStore(initialState) {
 	return createStore(rootReducer, initialState, enhancer);
 }
