+++ task-2/app/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { render } from 'react-dom';
 import { AppContainer } from 'react-hot-loader';
-import Root from './containers/Root';
-import { configureStore, history } from './store/configureStore';
-import './app.global.css';
+import Root from './views/Root';
+import { configureStore, history } from './data/store/configureStore';
+import './assets/app.global.css';
 
 const store = configureStore();
 
@@ -15,8 +15,8 @@ render(
 );
 
 if (module.hot) {
-	module.hot.accept('./containers/Root', () => {
-		const NextRoot = require('./containers/Root'); // eslint-disable-line global-require
+	module.hot.accept('./views/Root', () => {
+		const NextRoot = require('./views/Root'); // eslint-disable-line global-require
 		render(
 			<AppContainer>
 				<NextRoot store={store} history={history} />
