--- task-2/app/reducers/counter.js
@@ -1,13 +1,0 @@
-import { DECREMENT_COUNTER, INCREMENT_COUNTER } from '../actions/counter';
-import type { Action } from './types';
-
-export default function counter(state: number = 0, action: Action) {
-	switch (action.type) {
-		case INCREMENT_COUNTER:
-			return state + 1;
-		case DECREMENT_COUNTER:
-			return state - 1;
-		default:
-			return state;
-	}
-}
