--- task-2/app/reducers/index.js
@@ -1,11 +1,0 @@
-import { combineReducers } from 'redux';
-import { connectRouter } from 'connected-react-router';
-import counter from './counter';
-
-export default function createRootReducer(history: {}) {
-	const routerReducer = connectRouter(history)(() => {});
-
-	return connectRouter(history)(
-		combineReducers({ router: routerReducer, counter })
-	);
-}
