--- task-2/app/reducers/types.js
@@ -1,15 +1,0 @@
-import type { Dispatch as ReduxDispatch, Store as ReduxStore } from 'redux';
-
-export type counterStateType = {
-	+counter: number
-};
-
-export type Action = {
-	+type: string
-};
-
-export type GetState = () => counterStateType;
-
-export type Dispatch = ReduxDispatch<Action>;
-
-export type Store = ReduxStore<GetState, Action>;
