+++ task-2/app/views/App.js
