+++ task-2/test/containers/CounterPage.spec.js
@@ -4,8 +4,8 @@ import Adapter from 'enzyme-adapter-react-16';
 import { Provider } from 'react-redux';
 import { createBrowserHistory } from 'history';
 import { ConnectedRouter } from 'connected-react-router';
-import CounterPage from '../../app/containers/CounterPage';
-import { configureStore } from '../../app/store/configureStore';
+import CounterPage from '../../app/views/CounterPage';
+import { configureStore } from '../../app/data/store/configureStore';
 
 Enzyme.configure({ adapter: new Adapter() });
